tests: sphinx alignment unit testing port
The following ports over the Sphinx alignment tests from the legacy
container into the unit testing root.
