+++ tests/unit-tests/datasets/common/alignment.rst
@@ -1,6 +1,8 @@
+:orphan:
+
 default alignment
 -----------------
 
 .. figure:: https://example.com/image.png
 
-   caption
+    caption
