--- tests/unit-tests/legacy/common/expected/alignment-center.conf
@@ -1,8 +1,0 @@
-<ac:structured-macro ac:name="info">
-    <ac:parameter ac:name="icon">false</ac:parameter>
-    <ac:rich-text-body>
-        <ac:image ac:align="center">
-            <ri:url ri:value="https://example.com/image.png" />
-        </ac:image><p style="clear: both;text-align: center;">caption</p>
-    </ac:rich-text-body>
-</ac:structured-macro>
