--- tests/unit-tests/legacy/common/expected/alignment-left.conf
@@ -1,8 +1,0 @@
-<ac:structured-macro ac:name="info">
-    <ac:parameter ac:name="icon">false</ac:parameter>
-    <ac:rich-text-body>
-        <ac:image ac:align="left">
-            <ri:url ri:value="https://example.com/image.png" />
-        </ac:image><p style="clear: both;">caption</p>
-    </ac:rich-text-body>
-</ac:structured-macro>
