--- tests/unit-tests/legacy/common/expected/alignment-right.conf
@@ -1,8 +1,0 @@
-<ac:structured-macro ac:name="info">
-    <ac:parameter ac:name="icon">false</ac:parameter>
-    <ac:rich-text-body>
-        <ac:image ac:align="right" ac:style="float: right;">
-            <ri:url ri:value="https://example.com/image.png" />
-        </ac:image><p style="clear: both;text-align: right;">caption</p>
-    </ac:rich-text-body>
-</ac:structured-macro>
